# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of contacts app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://example.com/api/v1"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Operations with contacts."
    },
    {
      "name": "Auth",
      "description": "Auth operations."
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contacts",
        "operationId": "getAllContacts",
        "description": "Get list of contacts, filter and sort them",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "example": 10
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "example": "name",
            "description": "All contact`s fields can be mentioned. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "example": "asc",
            "description": "The default value is \"asc\""
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string",
              "default": ""
            },
            "example": "Ivan",
            "description": "Filter contacts by name, including symbols in the query"
          },
          {
            "in": "query",
            "name": "phoneNumber",
            "schema": {
              "type": "string",
              "default": ""
            },
            "example": "1819",
            "description": "Filter contacts by phone number, including symbols in the query"
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string",
              "default": ""
            },
            "example": "ivan_durak@example.com",
            "description": "Filter contacts by email, including symbols in the query"
          },
          {
            "in": "query",
            "name": "isFavorite",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "example": true,
            "description": "Filter contacts by favorite field"
          }
        ],
        "responses": {
          "200": {
            "description": "The contacts have been successfully found!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "The contacts have been successfully found!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contact"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 50
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 20
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create a new contact",
        "operationId": "addNewContact",
        "description": "Create a contact with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber"
                ],
                "properties": {
                  "name": {
                    "description": "Contact`s name",
                    "type": "string",
                    "example": "Ivan Durak"
                  },
                  "phoneNumber": {
                    "description": "Contact`s phone number",
                    "type": "string",
                    "example": "380999999"
                  },
                  "email": {
                    "description": "Contact`s email",
                    "type": "string",
                    "example": "ivan_durak@example.com"
                  },
                  "isFavourite": {
                    "description": "Is contact favorite",
                    "type": "boolean",
                    "default": false,
                    "example": true
                  },
                  "contactType": {
                    "description": "Contact`s type",
                    "type": "string",
                    "enum": [
                      "home",
                      "personal"
                    ],
                    "default": "personal",
                    "example": "home"
                  },
                  "photo": {
                    "description": "Contact`s photograph link",
                    "type": "string",
                    "example": "http://localhost:3000/auth/uploads/1719196919385_26702.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a contact!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/contacts/{contactId}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contact by id",
        "operationId": "getContactById",
        "description": "Get contact by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get contact by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 66738712d7702921c0987282!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "userId",
          "phoneNumber"
        ],
        "properties": {
          "_id": {
            "description": "Contact`s id",
            "type": "string",
            "example": "66738712d7702921c0987282"
          },
          "name": {
            "description": "Contact`s name",
            "type": "string",
            "example": "Ivan Durak"
          },
          "userId": {
            "description": "UserId",
            "type": "string",
            "example": "666f9fbc6c7a6b62825a3c68"
          },
          "phoneNumber": {
            "description": "Contact`s phone number",
            "type": "string",
            "example": "380999999"
          },
          "email": {
            "description": "Contact`s email",
            "type": "string",
            "example": "ivan_durak@example.com"
          },
          "isFavorite": {
            "description": "Is contact favorite",
            "type": "boolean",
            "example": true
          },
          "contactType": {
            "description": "Contact`s type",
            "type": "string",
            "enum": [
              "home",
              "personal"
            ],
            "example": "work"
          },
          "photo": {
            "description": "Contact`s photograph link",
            "type": "string",
            "example": "http://localhost:3000/auth/uploads/1719196919385_26702.jpg"
          },
          "createdAt": {
            "description": "Contact`s created data",
            "type": "string",
            "example": "2024-06-21T03:02:44.125+00:00"
          },
          "updatedAt": {
            "description": "Contact`s updated data",
            "type": "string",
            "example": "2024-06-24T02:41:59.399+00:00"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Please provide the Authorization header"
                },
                "data": {
                  "type": "object",
                  "example": {}
                }
              }
            }
          }
        }
      }
    }
  }
}